+++ NEWS.md
@@ -1,3 +1,3 @@
 # esaps 0.2.2
 
-* ...
+* add `psns()`, `dispro()` and `convert_esaps()`